requirements/requirements-core.txt
Django==4.2.16
python-dotenv==1.0.0
django-split-settings==1.2.0
daphne==4.1.2
django-comment-migrate==0.1.7
better_exceptions==0.3.3
whitenoise==6.7.0
loguru==0.7.2
singleton-decorator==1.0.0
wrapt==1.16.0
requests==2.31.0
django-debug-toolbar==4.2.0
psycopg2-binary==2.9.9
django-cors-headers==4.2.0

djangorestframework==3.14.0
django-filter==23.2
djangorestframework_simplejwt==5.3.1
drf-yasg==1.21.7

django-celery-beat==2.5.0
django-celery-results==2.5.0
sqlalchemy==2.0.20
flower==2.0.1

django-unfold==0.39.0
django-select2==8.1.2
django-simple-history==3.5.0
django-minio-backend==3.6.0
django-import-export[all]==4.0.7

django-auditlog==2.3.0
pydantic==2.5.3
django-guardian==2.4.0

jinja2==3.1.4
python-keycloak==3.7.0
kubernetes==29.0.0